+++ src/screens/Mnemonic.tsx
@@ -3,11 +3,11 @@ import useCopy from '@comps/hooks/Copy'
 import { ExclamationIcon } from '@comps/Icons'
 import Txt from '@comps/Txt'
 import { isIOS } from '@consts'
-import { generateMnemonic } from '@db/backup'
 import type { IMnemonicPageProps } from '@model/nav'
 import { useThemeContext } from '@src/context/Theme'
 import { NS } from '@src/i18n'
 import { getPinpadBg, mainColors } from '@styles'
+import { generateMnemonic } from '@wallet/restore'
 import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { FlatList, SafeAreaView, View } from 'react-native'
