+++ src/screens/Recovering.tsx
@@ -3,8 +3,8 @@ import Txt from '@comps/Txt'
 import type { IRecoveringPageProps } from '@model/nav'
 import { useThemeContext } from '@src/context/Theme'
 import { l } from '@src/logger'
-import { restoreWallet } from '@src/wallet'
 import { globals } from '@styles'
+import { restoreWallet } from '@wallet/restore'
 import { useEffect } from 'react'
 import { View } from 'react-native'
 import { s, ScaledSheet } from 'react-native-size-matters'
